events: GenericEventsFromChanges supports deletes
