+++ nomad/state/events.go
@@ -62,6 +62,7 @@ type JobDrainDetails struct {
 
 var MsgTypeEvents = map[structs.MessageType]string{
 	structs.NodeRegisterRequestType:                 TypeNodeRegistration,
+	structs.NodeDeregisterRequestType:               TypeNodeDeregistration,
 	structs.UpsertNodeEventsType:                    TypeNodeEvent,
 	structs.EvalUpdateRequestType:                   TypeEvalUpdated,
 	structs.AllocClientUpdateRequestType:            TypeAllocUpdated,
@@ -81,7 +82,7 @@ var MsgTypeEvents = map[structs.MessageType]string{
 }
 
 // GenericEventsFromChanges returns a set of events for a given set of
-// transaction changes. It currently ignores Delete operations.
+// transaction changes.
 func GenericEventsFromChanges(tx ReadTxn, changes Changes) (*structs.Events, error) {
 	eventType, ok := MsgTypeEvents[changes.MsgType]
 	if !ok {
@@ -90,10 +91,6 @@ func GenericEventsFromChanges(tx ReadTxn, changes Changes) (*structs.Events, err
 
 	var events []structs.Event
 	for _, change := range changes.Changes {
-		if change.Deleted() {
-			continue
-		}
-
 		if event, ok := eventFromChange(change); ok {
 			event.Type = eventType
 			event.Index = changes.Index
@@ -105,6 +102,21 @@ func GenericEventsFromChanges(tx ReadTxn, changes Changes) (*structs.Events, err
 }
 
 func eventFromChange(change memdb.Change) (structs.Event, bool) {
+	if change.Deleted() {
+		switch before := change.Before.(type) {
+		case *structs.Node:
+			return structs.Event{
+				Topic: structs.TopicNode,
+				Key:   before.ID,
+				Payload: &NodeEvent{
+					Node: before,
+				},
+			}, true
+		}
+
+		return structs.Event{}, false
+	}
+
 	switch after := change.After.(type) {
 	case *structs.Evaluation:
 		return structs.Event{
